requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-admin==2.0.1
django-celery-beat==2.2.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-rest-framework==0.1.0
django-six==1.0.4
django-timezone-field==4.1.2
djangorestframework==3.12.4
idna==2.10
kombu==5.1.0
prompt-toolkit==3.0.19
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
xlwt==1.3.0
